Add acceptance test for rover navigation (turn left)
Switched from TDDing with unit tests to acceptance tests. This "outside
in" approach was to speed up development as I was getting bogged down in
the details with the unit tests. Going outside in, you can focus on
getting the desired behaviour working first and later refactor while
keeping the tests green. Unit tests still need to be added, mostly for
non-critical sad paths which acceptance tests do not focus on.

Note that the feature was written to be high level enough to not be tied
to the implementation. Basically, if the implementation changes, you
should not have to change the wording in the feature.