Add integration test for Rover not moving
Was debating adding this. Is it a critical sad path?

Can be argued that it is. If a Rover were to somehow be fed an empty
list of instructions, you would expect it to not do anything, besides
logging/alerting. It would be an expensive mistake if it were to execute
the SelfDestruct command for example.

This test could be pushed lower down the testing pyramid as a unit test.
We could have a message expectation on an abstract base class `Command`
for example, but that pattern isn't necessary in a dynamic language like
Ruby and changing the design for a simple test is not worth it. We also
do not want to stub the system under test (which would be another
option).