:green_heart: Rover can move horizontally
The Plateau has moved from a composite object belonging to the Rover,
to belonging to the (Rover's) Position. This is because a Position
has no meaning without context - a position has to live within a "world"
which in this case is the Plateau that defines the grid of the world.